Added an InputManager to handle player input, and use one in the GameManager.
Also changed the game loop to render a frame first, then get input, and finally process the game frame, so now the rendering for a frame occurs in the frame after the processing.

Finally, fixed the GameManager default constructor so that it is actually the default constructor, not with args.